refactor(core): split facet panel layout into geometry and place modules
Extract grid geometry (size, margins, col/row) from placement packing
and the chrome entrypoint. Keep a thin barrel so import paths stay
stable, share FacetGridLayoutInput across the seam, and pin
placeFacetPanelsFromChrome with a characterization test.
